+++ libsensor/source/arputil.hpp
@@ -1,14 +1,13 @@
 #ifndef ARPUTIL_H_
 #define ARPUTIL_H_
 
-#include <stdlib.h>
-#include <stdint.h>
-#include <stdbool.h>
 
 #include <netinet/ether.h>
 #include <netinet/in.h>
 
-#include "netinfo.h"
+#include "netinfo.hpp"
+
+
 
 /* Protocol types */
 struct arp_ip4 {
@@ -23,7 +22,6 @@ struct arp_ip4 {
 void arp_reply_create(uint8_t buffer[ARP_IP4_SIZE], uint32_t my_ip4, const uint8_t my_hw[ETH_ALEN], uint32_t to_ip4, const uint8_t to_hw[ETH_ALEN]);
 void arp_request_create(uint8_t buffer[ARP_IP4_SIZE], uint32_t my_ip4, const uint8_t my_hw[ETH_ALEN], uint32_t to_ip4);
 void arp_request_set_to_ip(uint8_t *buffer, uint32_t to_ip4);
-bool arp_is_reply(const uint8_t *buffer, int length, const struct InterfaceInfo *current, struct NetAddress *out);
-
+bool arp_is_reply(const uint8_t *buffer, size_t length, const struct InterfaceInfo *current, struct NetAddress *out);
 
 #endif /* ARPUTIL_H_ */
